pyproject.toml
[project]
name = "ogboost"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
